+++ NEWS.md
@@ -1,3 +1,6 @@
+## Changes since last release
+- Update `load_forecasts_repo()` and `load_latest_forecasts_repo()` to use local data objects to validate `targets` parameter.
+
 ## covidHubUtils 0.1.5
 
 This is a release for renaming `plot_forecast()` to `plot_forecasts()`. `plot_forecast()` is still available to use but will return deprecation warnings to the user.